- создает новый индивидуальный презентер для точки маршрута
